Rename files to be more general
