+++ Makefile
@@ -12,8 +12,8 @@ CLANG_FORMAT=clang-format -i
 all: build
 	cmake --build build
 
-build: CMakeLists.txt test/CMakeLists.txt cmd/CMakeLists.txt
-	cmake -Bbuild -DCMAKE_BUILD_TYPE=Debug -DQMEDIA_BUILD_TESTS=ON -DBUILD_TESTING=ON  .
+${BUILD_DIR}: CMakeLists.txt
+	cmake -B${BUILD_DIR} -DCMAKE_BUILD_TYPE=Debug -DQMEDIA_BUILD_TESTS=ON -DBUILD_TESTING=ON  .
 
 clean:
 	cmake --build build --target clean
